User Authentication Done and Backend-API Completed

+++ controllers/contactController.js
@@ -3,15 +3,15 @@ const Contact = require("../models/conatctModel");
 
 // @desc Get all contacts
 // @route GET /api/contacts
-// @access public
+// @access private
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.user._id });
   res.status(200).json(contacts);
 });
 
 // @desc Create new contacts
 // @route POST /api/contacts
-// @access public
+// @access private
 const createContact = asyncHandler(async (req, res) => {
   console.log("The request body is :", req.body);
   const { name, email, phone } = req.body;
@@ -19,13 +19,13 @@ const createContact = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("All fields are mandatory!");
   }
-  const contact = await Contact.create({ name, email, phone });
+  const contact = await Contact.create({ name, email, phone, user_id: req.user._id, });
   res.status(201).json(contact);
 });
 
 // @desc Get individual contact
 // @route GET /api/contacts/:id
-// @access public
+// @access private
 const getContact = asyncHandler(async (req, res) => {
   try {
     const contact = await Contact.findById(req.params.id);
@@ -38,38 +38,46 @@ const getContact = asyncHandler(async (req, res) => {
 
 // @desc Update a contact
 // @route PUT /api/contacts/:id
-// @access public
+// @access private
 const updateContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("Some Error Occuerred");
   }
+  if (contact.user_id.toString() !== req.user._id) {
+    res.status(403);
+    throw new Error("User don't have permission to update other user contacts");
+  }
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
     req.body,
     { new: true }
-  );
-  res.status(200).json(updatedContact);
-});
-
-// @desc Delete a contact
-// @route DELETE /api/contacts/:id
-// @access public
-const deleteContact = asyncHandler(async (req, res) => {
-  const contact = await Contact.findById(req.params.id);
-  if (!contact) {
-    res.status(404);
-    throw new Error("Some Error Occuerred");
-  }
-  await Contact.deleteOne({ _id: contact._id });
-  res.status(200).json(contact);
-});
-
-module.exports = {
-  getContacts,
-  createContact,
-  getContact,
-  updateContact,
-  deleteContact,
-};
+    );
+    res.status(200).json(updatedContact);
+  });
+  
+  // @desc Delete a contact
+  // @route DELETE /api/contacts/:id
+  // @access private
+  const deleteContact = asyncHandler(async (req, res) => {
+    const contact = await Contact.findById(req.params.id);
+    if (!contact) {
+      res.status(404);
+      throw new Error("Some Error Occuerred");
+    }
+    if (contact.user_id.toString() !== req.user._id) {
+      res.status(403);
+      throw new Error("User don't have permission to update other user contacts");
+    }
+    await Contact.deleteOne({ _id: contact._id });
+    res.status(200).json(contact);
+  });
+  
+  module.exports = {
+    getContacts,
+    createContact,
+    getContact,
+    updateContact,
+    deleteContact,
+  };
